ISSUE #5244 address feedback
